server/beacon_server/requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0

# Database Drivers
mysqlclient==2.2.0
PyMySQL==1.1.0

# Environment Variables
python-dotenv==1.0.0

# Image Processing (for user avatars/media)
Pillow==10.1.0

# Security & Authentication
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# API Documentation
drf-yasg==1.21.7

# Form Handling & Validation
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Timezone Support
pytz==2023.3

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production Dependencies
gunicorn==21.2.0
whitenoise==6.6.0

# Caching (optional)
redis==5.0.1
django-redis==5.4.0

# Celery for Background Tasks (optional)
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Email Backend (optional)
django-sendgrid-v5==1.2.3

# Monitoring & Logging
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Static Files & Storage
django-storages==1.14.2
boto3==1.34.0

# Data Validation
marshmallow==3.20.1
django-filter==23.5

# Utilities
requests==2.31.0
python-decouple==3.8
Faker==20.1.0

# Performance
django-cachalot==2.6.1

# API Rate Limiting
django-ratelimit==4.1.0

# Admin Interface Enhancements
django-admin-interface==0.26.1
django-colorfield==0.10.1